Route Admin secure

+++ controller/controllerAdmin.mjs
@@ -13,9 +13,8 @@ const PRIVATE_KEY = "XxXMinecraftXxXBlackGaming"
 
 const verifyToken = (token) => {
     if (!token) {
-      return {message: "A token is required for authentification"};
+      return {message: "A token is required"};
     }
-
     try {
       return jwt.verify(token, PRIVATE_KEY);
     } catch (err) {
@@ -24,42 +23,62 @@ const verifyToken = (token) => {
 };
 
 export const adminController = {
-    findByLogin : async (login) => {
-    try {
-        return await adminDao.findByLogin(login)
-    } catch(e) { return Promise.reject({message: "error"})}
-    },
-    deleteByLogin: async (login) =>{
+
+    findAll : async (token) => {
         try {
+            if (!verifyToken(token).login) {
+                return Promise.reject({message: "not found"})
+            }
+            return await adminDao.findAll()
+        } catch(e) {
+            return Promise.reject({message: "error"})
+        }
+    },
+
+    deleteByLogin: async (login,token) =>{
+        try {
+            if (!verifyToken(token).login) {
+                return Promise.reject({message: "not found"})
+            }
             return await adminDao.deleteByLogin(login)
-        } catch(e)
-        { return Promise.reject({message: "error"})}
+        } catch(e){
+             return Promise.reject({message: "error"})
+        }
     },
-    update: async (login, admin) => {
+
+    update: async (login,admin,token) => {
         try {
+            if (!verifyToken(token).login) {
+                return Promise.reject({message: "not found"})
+            }
             admin.password= bcrypt.hashSync(admin.password,8)
             return await adminDao.update(login, admin)
         } catch (e) {
             return Promise.reject({message: "error"})
         }
     },
-    add:async (admin) => {
+
+    add:async (admin,token) => {
         try {
+            if (!verifyToken(token).login) {
+                return Promise.reject({message: "not found"})
+            }
             return  await adminDao.add(new Admin({
                 login : admin.login,
                 password : bcrypt.hashSync(admin.password,8),
                 token : ""
             }))
         } catch(e) {
-            return Promise.reject({message: "error"})}
+            return Promise.reject({message: "error"})
+        }
     },
+
     login : async (admin) => {
         try {
             const adminFound = await adminDao.findByLogin(admin.login)
             if (adminFound==null || !bcrypt.compareSync(admin.password, adminFound.password)) {
                 return {message: "not found"}
             }
-
             // on crée un token permettant à l'utilisateur de rejouer ce token quand il veut réaliser une action en lien avec son compte
             const token = jwt.sign(
                 { login: admin.login },
